fix deadlock
