+++ crates/strategy/zksync/src/executor.rs
@@ -1,29 +1,43 @@
 use std::sync::{Arc, Mutex};
 
 use alloy_primitives::{Address, U256};
+use alloy_rpc_types::serde_helpers::OtherFields;
 use foundry_cheatcodes::strategy::CheatcodeInspectorStrategyExt;
 
 use foundry_evm::{
-    backend::{strategy::BackendStrategyExt, BackendResult},
+    backend::{strategy::BackendStrategyExt, BackendResult, DatabaseExt},
     executors::{
         strategy::{EvmExecutorStrategy, ExecutorStrategy, ExecutorStrategyExt},
         Executor,
     },
+    InspectorExt,
 };
 use foundry_zksync_compiler::DualCompiledContracts;
-use revm::Database;
+use foundry_zksync_core::ZkTransactionMetadata;
+use revm::{
+    primitives::{EnvWithHandlerCfg, HashMap, ResultAndState},
+    Database,
+};
+use zksync_types::H256;
 
 use crate::{ZksyncBackendStrategy, ZksyncCheatcodeInspectorStrategy};
 
 #[derive(Debug, Default, Clone)]
 pub struct ZksyncExecutorStrategy {
     evm: EvmExecutorStrategy,
+    inspect_context: Option<ZkTransactionMetadata>,
+    persisted_factory_deps: HashMap<H256, Vec<u8>>,
     dual_compiled_contracts: DualCompiledContracts,
 }
 
 impl ExecutorStrategy for ZksyncExecutorStrategy {
     fn name(&self) -> &'static str {
         "zk"
+    }
+
+    fn set_inspect_context(&mut self, other_fields: OtherFields) {
+        let maybe_context = get_zksync_transaction_metadata(&other_fields);
+        self.inspect_context = maybe_context;
     }
 
     fn set_balance(
@@ -68,6 +82,50 @@ impl ExecutorStrategy for ZksyncExecutorStrategy {
             self.dual_compiled_contracts.clone(),
         )))
     }
+
+    fn call_inspect<'i, 'db>(
+        &mut self,
+        db: &'db mut dyn DatabaseExt,
+        env: &mut EnvWithHandlerCfg,
+        inspector: &'i mut dyn InspectorExt,
+    ) -> eyre::Result<ResultAndState> {
+        match self.inspect_context.take() {
+            None => self.evm.call_inspect(db, env, inspector),
+            Some(zk_tx) => foundry_zksync_core::vm::transact(
+                Some(&mut self.persisted_factory_deps),
+                Some(zk_tx.factory_deps),
+                zk_tx.paymaster_data,
+                env,
+                db,
+            ),
+        }
+    }
+
+    fn transact_inspect<'i, 'db>(
+        &mut self,
+        db: &'db mut dyn DatabaseExt,
+        env: &mut EnvWithHandlerCfg,
+        executor_env: &EnvWithHandlerCfg,
+        inspector: &'i mut dyn InspectorExt,
+    ) -> eyre::Result<ResultAndState> {
+        match self.inspect_context.take() {
+            None => self.evm.transact_inspect(db, env, executor_env, inspector),
+            Some(zk_tx) => {
+                // apply fork-related env instead of cheatcode handler
+                // since it won't be run inside zkvm
+                env.block = executor_env.block.clone();
+                env.tx.gas_price = executor_env.tx.gas_price;
+
+                foundry_zksync_core::vm::transact(
+                    Some(&mut self.persisted_factory_deps),
+                    Some(zk_tx.factory_deps),
+                    zk_tx.paymaster_data,
+                    env,
+                    db,
+                )
+            }
+        }
+    }
 }
 
 impl ExecutorStrategyExt for ZksyncExecutorStrategy {
@@ -77,4 +135,11 @@ impl ExecutorStrategyExt for ZksyncExecutorStrategy {
     ) {
         self.dual_compiled_contracts = dual_compiled_contracts;
     }
+}
+
+/// Retrieve metadata for zksync tx
+pub fn get_zksync_transaction_metadata(
+    other_fields: &OtherFields,
+) -> Option<ZkTransactionMetadata> {
+    other_fields.get_deserialized::<ZkTransactionMetadata>("zk_tx").transpose().ok().flatten()
 }
